# docs/openapi.yaml
openapi: 3.1.0
info:
  title: D1LANDANDHOUSE API
  version: 1.0.0
  description: |
    REST API สำหรับระบบเว็บไซต์นำเสนออสังหาริมทรัพย์ระดับ Luxury และระบบหลังบ้าน

    - `/api/v1/public/*`  — สำหรับเว็บไซต์หน้าบ้าน (ไม่ต้องล็อกอิน ยกเว้น `/public/me/*`)
    - `/api/v1/admin/*`   — สำหรับ Backoffice (ต้องล็อกอิน + RBAC: owner | admin)

    ระบบนี้เป็นเว็บนำเสนอทรัพย์และเก็บลีดเท่านั้น ไม่มีการซื้อขาย/ชำระเงินผ่าน API
  contact:
    name: Development Team
  license:
    name: Proprietary
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: https://stg-api.example.com/api/v1
    description: Staging
  - url: http://localhost:4000/api/v1
    description: Local
...
security: []
paths:
...
  /public/projects:
    get:
      tags: ["Public: Projects"]
      summary: รายการโครงการ
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - { name: projectType, in: query, schema: { type: string } }
        - { name: constructionStatus, in: query, schema: { type: string } }
        - { name: provinceId, in: query, schema: { type: string } }
        - { name: keyword, in: query, schema: { type: string } }
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/ProjectListItem' } }
  /public/projects/{slug}:
    get:
      tags: ["Public: Projects"]
      summary: รายละเอียดโครงการ
      parameters: [ { $ref: '#/components/parameters/Slug' } ]
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/ProjectDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /public/leads:
    post:
      tags: ["Public: Leads"]
      summary: ส่งฟอร์มติดต่อ (ทรัพย์ / ทั่วไป / floating)
      description: |
        ป้องกันสแปมด้วย reCAPTCHA v3 + rate limit 5 ครั้ง/10 นาที/IP + honeypot
        ถ้าเบอร์โทรซ้ำกับลีดเดิมของทรัพย์เดียวกันภายใน 24 ชม. ระบบจะบันทึกเป็น activity
        ในลีดเดิมแทนการสร้างลีดใหม่
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LeadCreateRequest' }
      responses:
        '201':
          description: รับลีดแล้ว
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      refNo: { type: string, example: LD-20260914-0042 }
                      message: { $ref: '#/components/schemas/LocalizedString' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /admin/users:
    get:
      tags: ["Admin: Users"]
      summary: รายการผู้ใช้ระบบ
      description: '**เฉพาะ Owner** — Admin เรียกจะได้ 403'
      security: [ { adminAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - { name: role, in: query, schema: { type: string, enum: [owner, admin] } }
        - { name: status, in: query, schema: { type: string, enum: [active, suspended] } }
      responses:
        '200': { description: สำเร็จ }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: ["Admin: Users"]
      summary: สร้างบัญชีผู้ใช้ใหม่ (เฉพาะ Owner)
      security: [ { adminAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, role]
              properties:
                email: { type: string, format: email }
                name: { type: string }
                phone: { type: string }
                role: { type: string, enum: [owner, admin] }
                isAgent: { type: boolean }
      responses:
        '201': { description: สร้างแล้ว ส่งอีเมลตั้งรหัสผ่าน }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/users/{id}:
    patch:
      tags: ["Admin: Users"]
      summary: แก้ไขผู้ใช้ (เฉพาะ Owner)
      security: [ { adminAuth: [] } ]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content: { application/json: { schema: { type: object } } }
      responses:
        '200': { description: สำเร็จ }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      tags: ["Admin: Users"]
      summary: ลบผู้ใช้ (เฉพาะ Owner — ห้ามลบ Owner คนสุดท้าย)
      security: [ { adminAuth: [] } ]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses:
        '204': { description: ลบแล้ว }
        '409':
          description: ลบ Owner คนสุดท้ายไม่ได้ (code `LAST_OWNER`)
          content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }

  # ══════════════════════════ ADMIN: PROPERTIES ══════════════════════════
...
  /admin/media/{id}:
    patch:
      tags: ["Admin: Media"]
      summary: แก้ไข alt / caption / แท็ก
      security: [ { adminAuth: [] } ]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        content: { application/json: { schema: { type: object } } }
      responses: { '200': { description: สำเร็จ } }
    delete:
      tags: ["Admin: Media"]
      summary: ลบสื่อ
      security: [ { adminAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/Id'
        - { name: force, in: query, schema: { type: boolean }, description: ยืนยันลบแม้ยังถูกใช้งานอยู่ }
      responses:
        '204': { description: ลบแล้ว }
        '409': { description: 'สื่อถูกใช้งานอยู่ (code `MEDIA_IN_USE`)' }

  # ══════════════════════════ ADMIN: ARTICLES ══════════════════════════
...
  /admin/promotions:
    get:
      tags: ["Admin: Promotions"]
      summary: รายการโปรโมชั่น
      security: [ { adminAuth: [] } ]
      parameters:
        - { name: state, in: query, schema: { type: string, enum: [upcoming, active, expired, all], default: all } }
      responses: { '200': { description: สำเร็จ } }
    post:
      tags: ["Admin: Promotions"]
      summary: สร้างโปรโมชั่น
      security: [ { adminAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromotionUpsertRequest' }
      responses:
        '201': { description: สร้างแล้ว }
        '422':
          description: 'ช่วงเวลาไม่ถูกต้อง (code `INVALID_PROMOTION_PERIOD`)'
...
  /admin/leads:
    get:
      tags: ["Admin: Leads"]
      summary: ตารางลีด
      security: [ { adminAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - { name: status, in: query, schema: { type: string, enum: [new, contacted, viewing_scheduled, negotiating, won, lost] } }
        - { name: source, in: query, schema: { type: string } }
        - { name: assignedTo, in: query, schema: { type: string } }
        - { name: intent, in: query, schema: { type: string } }
        - { name: dateFrom, in: query, schema: { type: string, format: date } }
        - { name: dateTo, in: query, schema: { type: string, format: date } }
        - { name: q, in: query, schema: { type: string }, description: ค้นจากชื่อ/เบอร์/อีเมล/เลขอ้างอิง }
      responses: { '200': { description: สำเร็จ } }
    post:
      tags: ["Admin: Leads"]
      summary: สร้างลีดด้วยมือ (รับสายโทรศัพท์ / walk-in)
      security: [ { adminAuth: [] } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LeadCreateRequest' } } }
      responses: { '201': { description: สร้างแล้ว } }
...
  /admin/leads/{id}:
    get:
      tags: ["Admin: Leads"]
      summary: รายละเอียดลีด + ไทม์ไลน์
      security: [ { adminAuth: [] } ]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      responses: { '200': { description: สำเร็จ } }
  /admin/leads/{id}/assign:
    patch:
      tags: ["Admin: Leads"]
      summary: มอบหมายผู้ดูแลลีด
      security: [ { adminAuth: [] } ]
      parameters: [ { $ref: '#/components/parameters/Id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignedTo]
              properties: { assignedTo: { type: string } }
      responses: { '200': { description: สำเร็จ } }
...
components:
  securitySchemes:
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token ของพนักงาน (staff realm) — อายุ 15 นาที
    customerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token ของลูกค้าสมาชิก (customer realm)
  parameters:
    Page:   { name: page,  in: query, schema: { type: integer, default: 1, minimum: 1 } }
    Limit:  { name: limit, in: query, schema: { type: integer, default: 12, minimum: 1, maximum: 100 } }
    Id:     { name: id,    in: path,  required: true, schema: { type: string }, description: MongoDB ObjectId }
    Slug:   { name: slug,  in: path,  required: true, schema: { type: string } }
...
  responses:
...
    Forbidden:
      description: สิทธิ์ไม่พอ (เช่น Admin เรียก endpoint ของ Owner)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: ไม่พบข้อมูล
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Conflict:
      description: ข้อมูลขัดแย้ง
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    Unprocessable:
      description: ผิดกฎทางธุรกิจ
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    RateLimited:
      description: ส่งคำขอถี่เกินไป
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
  schemas:
    LocalizedString:
      type: object
      properties:
        th: { type: string }
        en: { type: string }
      required: [th]
    GeoPoint:
      type: object
      properties:
        type: { type: string, enum: [Point] }
        coordinates:
          type: array
          items: { type: number }
          minItems: 2
          maxItems: 2
          description: '[longitude, latitude] — ลำดับนี้เท่านั้น'
    MediaRef:
      type: object
      properties:
        mediaId: { type: string }
        url: { type: string, format: uri }
        variants:
          type: object
          properties:
            thumb: { type: string }
            medium: { type: string }
            large: { type: string }
            original: { type: string }
        alt: { $ref: '#/components/schemas/LocalizedString' }
        blurhash: { type: string }
        width: { type: integer }
        height: { type: integer }
        sortOrder: { type: integer }
    SeoMeta:
      type: object
      properties:
        metaTitle: { $ref: '#/components/schemas/LocalizedString' }
        metaDescription: { $ref: '#/components/schemas/LocalizedString' }
        ogImage: { $ref: '#/components/schemas/MediaRef' }
        keywords: { type: array, items: { type: string } }
        canonicalUrl: { type: string }
        noIndex: { type: boolean }
    PaginatedResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: array, items: {} }
        meta:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            totalPages: { type: integer }
            sort: { type: string }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string, example: ข้อมูลไม่ถูกต้อง }
            messageEn: { type: string, example: Validation failed }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
            requestId: { type: string }
...
    LocationInfo:
      type: object
      properties:
        address: { $ref: '#/components/schemas/LocalizedString' }
        provinceId: { type: string }
        districtId: { type: string }
        subdistrictId: { type: string }
        zone: { type: string }
        postalCode: { type: string }
        geo: { $ref: '#/components/schemas/GeoPoint' }
        hideExactLocation: { type: boolean }
    NearbyPlace:
      type: object
      properties:
        type: { type: string, enum: [bts, mrt, arl, school, university, hospital, mall, market, airport, expressway, park, office] }
        name: { $ref: '#/components/schemas/LocalizedString' }
        distanceKm: { type: number }
        travelMinutes: { type: integer }
    VideoRef:
      type: object
      properties:
        provider: { type: string, enum: [youtube, vimeo, file] }
        url: { type: string }
        videoId: { type: string }
        title: { $ref: '#/components/schemas/LocalizedString' }
        thumbnail: { type: string }
        sortOrder: { type: integer }
...
    ProjectListItem:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { $ref: '#/components/schemas/LocalizedString' }
        developer: { $ref: '#/components/schemas/LocalizedString' }
        projectType: { type: string }
        constructionStatus: { type: string }
        priceRange:
          type: object
          properties: { min: { type: number }, max: { type: number } }
        totalUnits: { type: integer }
        availableUnits: { type: integer }
        coverImage: { $ref: '#/components/schemas/MediaRef' }
        location: { $ref: '#/components/schemas/LocationInfo' }
    ProjectDetail:
      allOf:
        - $ref: '#/components/schemas/ProjectListItem'
        - type: object
          properties:
            description: { $ref: '#/components/schemas/LocalizedString' }
            completionDate: { type: string, format: date }
            totalBuildings: { type: integer }
            totalFloors: { type: integer }
            commonFee: { type: number }
            sinkingFund: { type: number }
            foreignQuotaAvailable: { type: boolean }
            unitTypes:
              type: array
              items:
                type: object
                properties:
                  name: { $ref: '#/components/schemas/LocalizedString' }
                  sizeSqmMin: { type: number }
                  sizeSqmMax: { type: number }
                  bedrooms: { type: integer }
                  priceFrom: { type: number }
                  floorPlan: { $ref: '#/components/schemas/MediaRef' }
            facilities: { type: array, items: { $ref: '#/components/schemas/Amenity' } }
            masterPlan: { type: array, items: { $ref: '#/components/schemas/MediaRef' } }
            gallery: { type: array, items: { $ref: '#/components/schemas/MediaRef' } }
            videos: { type: array, items: { $ref: '#/components/schemas/VideoRef' } }
            nearby: { type: array, items: { $ref: '#/components/schemas/NearbyPlace' } }
            seo: { $ref: '#/components/schemas/SeoMeta' }
...
    Amenity:
      type: object
      properties:
        id: { type: string }
        slug: { type: string }
        name: { $ref: '#/components/schemas/LocalizedString' }
        icon: { type: string }
        group: { type: string, enum: [unit, project, security, wellness, service] }
...
    PromotionUpsertRequest:
      type: object
      required: [title, discountType, startAt, endAt, appliesTo]
      properties:
        title: { $ref: '#/components/schemas/LocalizedString' }
        description: { $ref: '#/components/schemas/LocalizedString' }
        badgeText: { $ref: '#/components/schemas/LocalizedString' }
        badgeColor: { type: string, example: '#B3241E' }
        discountType: { type: string, enum: [percentage, fixed, custom] }
        discountValue: { type: number }
        maxDiscountAmount: { type: number }
        appliesTo: { type: string, enum: [all, properties, projects, property_types, locations] }
        targetPropertyIds: { type: array, items: { type: string } }
        targetProjectIds: { type: array, items: { type: string } }
        targetPropertyTypes: { type: array, items: { type: string } }
        targetLocationIds: { type: array, items: { type: string } }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        showCountdown: { type: boolean }
        priority: { type: integer }
        isActive: { type: boolean }
        terms: { $ref: '#/components/schemas/LocalizedString' }
    LeadCreateRequest:
      type: object
      required: [name, phone, intent, consentAccepted]
      properties:
        source: { type: string, enum: [property_form, general_form, floating_form, project_form, article_form, phone, line, walk_in, facebook, manual] }
        propertyId: { type: string, nullable: true }
        projectId: { type: string, nullable: true }
        promotionId: { type: string, nullable: true }
        name: { type: string, minLength: 2 }
        phone: { type: string, example: '0812345678' }
        email: { type: string, format: email }
        lineId: { type: string }
        preferredChannel: { type: string, enum: [phone, line, email, whatsapp, messenger] }
        preferredTime: { type: string, enum: [morning, afternoon, evening, anytime] }
        intent: { type: string, enum: [buy, rent, sell, invest, consult] }
        message: { type: string, maxLength: 2000 }
        budget:
          type: object
          properties: { min: { type: number }, max: { type: number } }
        interestedTypes: { type: array, items: { type: string } }
        interestedZones: { type: array, items: { type: string } }
        consentAccepted: { type: boolean, description: ต้องเป็น true ตาม PDPA }
        recaptchaToken: { type: string }
        pageUrl: { type: string }
        utm:
          type: object
          properties:
            source: { type: string }
            medium: { type: string }
            campaign: { type: string }
            term: { type: string }
            content: { type: string }
...